Clean few component
